aggiunto funzione azzeraConto

+++ app.js
@@ -6,16 +6,17 @@ class Account {
     constructor() {
         this.accounts = this.caricaAccounts();
         console.log(`array degli accounts: ${this.accounts}`);
+        this.creaCard(this.accounts);
     }
 
 
-    caricaAccounts(){
+    caricaAccounts() {
         const acc = JSON.parse(localStorage.getItem("accounts")) ?? [];
 
         return acc
     }
 
-    cercaAccount(arr, nome){
+    cercaAccount(arr, nome) {
         // console.log(`lunghezza: ${arr.length}, nome: ${nome}`);
         let temp = 0;
         for (let i = 0; i < arr.length; i++) {
@@ -30,9 +31,9 @@ class Account {
     }
 
     entrataUscita(nome, denaro) {
-        console.log("accounts------");
-        console.log(this.accounts);
-        console.log("---------");
+        // console.log("accounts------");
+        // console.log(this.accounts);
+        // console.log("---------");
         let presenza = this.cercaAccount(this.accounts, nome);
         // console.log(presenza);
 
@@ -46,7 +47,7 @@ class Account {
                 "nome": nome,
                 "bilancio": denaro
             };
-    
+
             let arrTemp = [temp];
 
             this.accounts.push(temp);
@@ -56,7 +57,7 @@ class Account {
         else {
             for (let i = 0; i < this.accounts.length; i++) {
                 if (nome == this.accounts[i]["nome"]) {
-                    
+
                     let denaroTemp = this.accounts[i]["bilancio"] += denaro;
                     //console.log("il denaro sul conto è: " + this.accounts[i]["bilancio"] + " il denaro da aggiungere " + denaro);
 
@@ -74,7 +75,7 @@ class Account {
                 }
             }
 
-            
+
         }
 
         // console.log("ehi ");
@@ -87,23 +88,37 @@ class Account {
 
         const seleziona = document.querySelector(".card-container");
 
+
+
         arr.forEach((item) => {
 
             let contenuto = `<div class="card-css">
-        <h5>${item.nome}</h5>
-        <p id="${item.nome}">${item.bilancio}</h5>
+            <h5>${item.nome}</h5>
+            <p id="${item.nome}">${item.bilancio}</h5>
         
-        </div>`;
+            </div>`;
 
             seleziona.innerHTML += contenuto;
         })
     }
 
-    modificaCard() {
+    azzeraConto(nome) {
+        if (confirm("azzerare il conto?")) {
+            for (let i = 0; i < this.accounts.length; i++) {
+                //console.log("ciao");
+                if (this.accounts[i]["nome"] == nome) {
+                    this.accounts[i]["bilancio"] = 0;
+                    console.log(this.accounts);
+                    localStorage.setItem("accounts", JSON.stringify(this.accounts));
+                    document.getElementById(nome).innerHTML = this.accounts[i]["bilancio"];
+                }
+            }
+        }
 
+        //console.log("ciao");
     }
 
-    aggiungiAccount(account){
+    aggiungiAccount(account) {
         const seleziona = document.getElementById("acc");
 
         let aggiugni = `<option value="${account}">${account}</option>`;
@@ -111,11 +126,14 @@ class Account {
         seleziona.innerHTML += aggiugni;
     }
 
-    ripulisci(){
-        this.accounts = [];
-        console.log(this.accounts);
-        document.querySelector(".card-container").innerHTML = "";
-        localStorage.clear();
+    ripulisci() {
+        if (confirm("ripulire tutto?")) {
+            //this.accounts = [];
+            console.log(this.accounts);
+            document.querySelector(".card-container").innerHTML = "";
+            localStorage.clear();
+        }
+
     }
 
 
@@ -129,9 +147,9 @@ let operazioni = new Account();
 
 
 
-document.getElementById("aggiungiAccount")?.addEventListener("click", () => {
+document.getElementById("aggiungiAccount").addEventListener("click", () => {
     let temp = document.getElementById("accountInput").value;
-    
+
     operazioni.aggiungiAccount(temp);
 })
 
@@ -145,10 +163,19 @@ function numero(stringa) {
     return pattern.test(stringa);
 }
 
+document.querySelectorAll(".card-css").forEach(item => {
+    item.addEventListener("click", (e) => {
+        const temp = e.currentTarget.querySelector("p").id;
+        const stringTemp = String(temp);
+        operazioni.azzeraConto(stringTemp);
+        //console.log(temp);
+    })
+})
+
 
 
 let bottone = document.getElementById("invia");
-bottone?.addEventListener("click", (e) => {
+bottone.addEventListener("click", (e) => {
 
     //let opzione: string = (document.getElementById("selezione") as HTMLSelectElement).value;
     let acc = (document.getElementById("acc")).value;
@@ -170,6 +197,6 @@ bottone?.addEventListener("click", (e) => {
 })
 
 
-document.getElementById("ripulisci")?.addEventListener("click", () => {
+document.getElementById("ripulisci").addEventListener("click", () => {
     operazioni.ripulisci();
 })
